backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
pydantic==2.9.2
pdfplumber==0.11.4
openai==1.58.1
python-dotenv==1.0.1
python-multipart==0.0.12
httpx==0.27.2
aiofiles==24.1.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
alembic==1.13.1
slowapi==0.1.9
reportlab==4.1.0
